requirements.txt
Babel==2.15.0
DataProperty==1.0.1
GitPython==3.1.43
Jinja2==3.1.4
MarkupSafe==2.1.5
PyYAML==6.0.1
Pygments==2.18.0
QtPy==2.4.1
Send2Trash==1.8.3
absl-py==2.1.0
accelerate==0.32.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi-bindings==21.2.0
argon2-cffi==23.1.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
autoawq==0.2.6
autoawq_kernels==0.0.7
beautifulsoup4==4.12.3
bitsandbytes==0.43.3
bleach==6.1.0
blobfile==2.1.1
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
datasets==2.20.0
debugpy==1.8.2
decorator==5.1.1
deepspeed==0.14.5
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
evaluate==0.4.2
exceptiongroup==1.2.2
executing==2.0.1
fairscale==0.4.13
fastjsonschema==2.20.0
filelock==3.15.4
# flash-attn==2.6.1
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.5
idna==3.7
immutables==0.20
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
jiter==0.5.0
joblib==1.4.2
json5==0.9.25
json_strong_typing==0.3.4
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema-specifications==2023.12.1
jsonschema==4.23.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter==1.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
llama_models==0.0.1
lm_eval==0.4.3
lxml==4.9.4
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
matplotlib==3.9.1
mbstrdecoder==1.1.3
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.0.2
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
notebook==7.2.1
notebook_shim==0.2.4
numexpr==2.10.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.555.43
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
openai==1.40.6
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathvalidate==3.2.0
peft==0.11.1
pexpect==4.9.0
pillow==10.4.0
pip==24.2
platformdirs==4.2.2
plotly==5.22.0
pluggy==1.5.0
portalocker==2.10.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.27.2
psutil==6.0.0
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow-hotfix==0.6
pyarrow==17.0.0
pybind11==2.13.1
pycparser==2.22
pycryptodomex==3.20.0
pydantic==1.10.13
pydantic_core==2.18.2
pyparsing==3.1.2
pytablewriter==1.2.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
pyzmq==26.0.3
qtconsole==5.5.2
quanto==0.2.0
ray==2.32.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rouge_score==0.1.2
rpds-py==0.19.0
sacrebleu==2.4.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentencepiece==0.2.0
sentry-sdk==2.10.0
setproctitle==1.2.2
# setproctitle==1.3.3
setuptools==71.0.0
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
sqlitedict==2.1.0
stack-data==0.6.3
sympy==1.13.0
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.6
tenacity==8.5.0
tensorboardX==2.6.2.2
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.7.0
tinyBenchmarks @ git+https://github.com/felipemaiapolo/tinyBenchmarks
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm-multiprocess==0.0.11
tqdm==4.66.4
traitlets==5.14.3
transformers-stream-generator==0.0.5
transformers==4.44.2
triton==2.3.1
trl==0.8.6
typepy==1.3.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tyro==0.8.5
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
wandb==0.17.4
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.43.0
widgetsnbextension==4.0.11
word2number==1.1
xxhash==3.4.1
yarl==1.9.4
zstandard==0.23.0
eth_keys==0.5.1
eth-hash[pycryptodome]==0.7.0
